+++ module.js
@@ -1,6 +1,6 @@
 module.exports = {
     extends: [
-        "eslint-config-konnektid/babel",
-        "eslint-config-konnektid/specs/module"
+        "eslint-config-konnektid/specs/module",
+        "eslint-config-konnektid/rules/module"
     ]
 };