+++ rules/module.js
@@ -1,0 +1,7 @@
+module.exports = {
+    rules: {
+
+        // ignore extra strict directives
+        "strict": 0
+    }
+};